Backport of arbitrary custom clevis pin in 3.5
In order to support new clevis pin, either they need to be added each
time in the hardcoded list of pins or ignition can allow any name for
the pin. This is required in order to enable the clevis trustee pin used
for confidential clusters.

The backport to 3.5 is necessary because the rust crate for ignition
only support up to 3.5 config version and cannot be used with
3.6-experimental.

Signed-off-by: Alice Frosi <afrosi@redhat.com>